refactor(ChangePlan): change components and add classnames library
