+++ package.json
@@ -19,6 +19,7 @@
     "axios": "^0.19.2",
     "axios-retry": "^3.1.8",
     "c3": "^0.7.18",
+    "classnames": "^2.2.6",
     "cross-env": "^7.0.2",
     "eclint": "^2.8.1",
     "env-cmd": "^10.1.0",
