--- src/components/ChangePlan/Card/Card.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Card = ({ children, cssClass }) => {
-  return <div class={`dp-card ${cssClass}`}>{children}</div>;
-};
-
-export default Card;
