--- src/components/ChangePlan/Card/Ribbon.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Ribbon = ({ children, position }) => {
-  return <div className={`dp-ribbon dp-ribbon-${position}`}>{children}</div>;
-};
-
-export default Ribbon;
