+++ src/components/ChangePlan/ChangePlan.js
@@ -2,10 +2,7 @@ import React from 'react';
 import { useIntl } from 'react-intl';
 import { useRouteMatch } from 'react-router-dom';
 import { Helmet } from 'react-helmet';
-import Card from './Card/Card';
-import CardPrice from './Card/CardPrice';
-import CardAction from './Card/CardAction';
-import Ribbon from './Card/Ribbon';
+import { Card, CardPrice, CardAction, Ribbon } from './Card';
 
 const ChangePlan = () => {
   const { params } = useRouteMatch();
@@ -39,27 +36,25 @@ const ChangePlan = () => {
       <div className="p-t-54 p-b-54" style={{ backgroundColor: '#f6f6f6', flex: '1' }}>
         <section className="dp-container">
           <div className="dp-rowflex">
-            <div className="col-sm-12" style={{ textAlign: 'center' }}>
+            <div className="dp-align-center p-t-30 p-b-30">
               <Card>
                 <div className="dp-content-plans">
                   <h3>{cardData.name}</h3>
                   <p>{_('change_plan.description')}</p>
                   <p>{cardData.descriptionPlan}</p>
-                  <CardPrice currency="US$">{cardData.price}</CardPrice>
-                  <CardAction url={cardData.action.url}>{cardData.action.text}</CardAction>
                 </div>
+                <CardPrice currency="US$">{cardData.price}</CardPrice>
+                <CardAction url={cardData.action.url}>{cardData.action.text}</CardAction>
               </Card>
-              <Card cssClass="dp-highlighthed">
-                <Ribbon position="top-right">
-                  <span>{_('change_plan.recommended')}</span>
-                </Ribbon>
+              <Card highlighted>
+                <Ribbon content={_('change_plan.recommended')} />
                 <div className="dp-content-plans">
                   <h3>{cardData.name}</h3>
                   <p>{_('change_plan.description')}</p>
                   <p>{cardData.descriptionPlan}</p>
-                  <CardPrice currency="US$">{cardData.price}</CardPrice>
-                  <CardAction url={cardData.action.url}>{cardData.action.text}</CardAction>
                 </div>
+                <CardPrice currency="US$">{cardData.price}</CardPrice>
+                <CardAction url={cardData.action.url}>{cardData.action.text}</CardAction>
               </Card>
             </div>
           </div>
